backend/myLubd/requirements.txt
# Core Django dependencies
Django>=3.2,<4.0
djangorestframework>=3.12.0
djangorestframework-simplejwt>=5.0.0
django-cors-headers>=3.10.0
django-filter>=2.4.0

# PDF Generation
reportlab>=3.6.0
Pillow>=8.0.0

# Database
psycopg2-binary>=2.9.0

# HTTP and API
requests>=2.25.0

# Google OAuth
google-auth>=2.0.0
google-auth-oauthlib>=1.0.0
google-auth-httplib2>=0.1.0

# Image processing
Pillow>=8.0.0

# Utilities
python-dateutil>=2.8.0

# Optional development dependencies
django-debug-toolbar>=4.0.0
django-dbbackup>=4.0.0
